textures_holder creado y andando
