+++ client/graphics.cpp
@@ -60,29 +60,7 @@ void Graphics::run(GameState *gs, Queue<std::string> &queue_comandos, Queue<Game
     }
 }
 GraphicsEntityHolder Graphics::start_graphics_entity(GameState *gs, SdlWindow &window) {
-    std::map<EntityType, std::map<AnimationName, std::shared_ptr<SdlTexture>>> textures_holder;
-
-    //quiero que todo este codigo este en otro lado
-
-    textures_holder[SOLDIER_IDF][AN_IDLE] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_IDF/Idle.png", window));
-    textures_holder[SOLDIER_IDF][AN_SHOOT] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_IDF/Shoot_1.png", window));
-    textures_holder[SOLDIER_IDF][AN_RUN] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_IDF/Run.png", window));
-    textures_holder[SOLDIER_IDF][AN_DIE] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_IDF/Dead.png", window));
-    textures_holder[SOLDIER_IDF][AN_RELOAD] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_IDF/Recharge.png", window));
-
-
-    textures_holder[SOLDIER_SCOUT][AN_IDLE] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_Scout/Idle.png", window));
-    textures_holder[SOLDIER_SCOUT][AN_SHOOT] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_Scout/Shoot_1.png", window));
-    textures_holder[SOLDIER_SCOUT][AN_RUN] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_Scout/Run.png", window));
-    textures_holder[SOLDIER_SCOUT][AN_DIE] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_Scout/Dead.png", window));
-    textures_holder[SOLDIER_SCOUT][AN_RELOAD] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_Scout/Recharge.png", window));
-
-    textures_holder[SOLDIER_P90][AN_IDLE] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_P90/Idle.png", window));
-    textures_holder[SOLDIER_P90][AN_SHOOT] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_P90/Shoot_1.png", window));
-    textures_holder[SOLDIER_P90][AN_RUN] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_P90/Run.png", window));
-    textures_holder[SOLDIER_P90][AN_DIE] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_P90/Dead.png", window));
-    textures_holder[SOLDIER_P90][AN_RELOAD] = std::shared_ptr<SdlTexture>(new SdlTexture("../../assets/Soldier_P90/Recharge.png", window));
-
+    TexturesHolder textures_holder(window);
     return GraphicsEntityHolder(gs, std::move(textures_holder), window);
 }
